Adding the first batch of changes to org-similarity.
These changes are on the backend of it. The python script prints the
list of files and not org links to those files, the function that runs
the script takes a node as an argument instead of defaulting to the
current buffer and then returns a list of files that are relevant to
that node with the node as the car of the list and there is another
function that takes this list and converts it into a list of
nodes. Then, using that list, all the magic can happen.